+++ packages/core/src/company-setting/index.ts
@@ -1,3 +1,2 @@
 export * from './create'
-export * from './fragments'
 export * from './types'