+++ packages/core/src/page/seo-meta/index.ts
@@ -1,2 +1,1 @@
-export * from './fragments'
 export * from './helpers'