+++ packages/core/tsdown.config.ts
@@ -6,8 +6,6 @@ export default defineConfig({
     './src/*/index.ts',
   ],
   copy: [
-    { from: 'src/image/components', to: 'dist/image' },
-    { from: 'src/page/components', to: 'dist/page' },
     { from: 'src/query-params/components', to: 'dist/query-params' },
   ],
   dts: true,